Python_GUI/requirements.txt
pyserial>=3.5
openpyxl>=3.1.0
pandas>=2.0.0
